refactor(server): rename the auth plugin module to resolve-user-id
