+++ packages/emitsignal-server/src/http/auth/resolve-user-id.ts
