+++ packages/emitsignal-server/src/http/billing/get.ts
@@ -2,7 +2,7 @@ import type { BillingInfo, PlanSubscriptionInfo } from '@emitsignal/shared';
 
 import Elysia from 'elysia';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
 import { prisma } from '#/lib/prisma';
 import { readAnonLimiter, readAuthLimiter } from '#/lib/rate-limit';
