+++ packages/emitsignal-server/src/http/messages/__tests__/attachments.spec.ts
@@ -8,7 +8,7 @@ mock.module('#/lib/prisma', () => ({ prisma: prismaMock }));
 mock.module('#/lib/storage', () => ({ FileStorageService: fileStorageMock }));
 
 const resolveUserIdMock = mock<() => Promise<null | string>>(() => Promise.resolve(null));
-mock.module('#/http/auth/plugin', () => ({ resolveUserId: resolveUserIdMock }));
+mock.module('#/http/auth/resolve-user-id', () => ({ resolveUserId: resolveUserIdMock }));
 
 import { attachments } from '#/http/messages/attachments';
 import { resetUserPlansForTests, setUserPlanForTests } from '#/services/billing/get-user-plan';
