+++ packages/emitsignal-server/src/http/push-tokens/register.ts
@@ -1,6 +1,6 @@
 import Elysia, { t } from 'elysia';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import { prisma } from '#/lib/prisma';
 
 export const registerPushToken = new Elysia({ prefix: '/push-tokens' }).post(
