+++ packages/emitsignal-server/src/http/push-tokens/update.ts
@@ -1,6 +1,6 @@
 import Elysia, { t } from 'elysia';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import { prisma } from '#/lib/prisma';
 
 export const updatePushToken = new Elysia({ prefix: '/push-tokens' }).patch(
