+++ packages/emitsignal-server/src/http/subscriptions/claim.ts
@@ -1,6 +1,6 @@
 import Elysia, { t } from 'elysia';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import { prisma } from '#/lib/prisma';
 
 export const claimSubscriptions = new Elysia({ prefix: '/subscriptions' }).post(
