+++ packages/emitsignal-server/src/http/topic/claim.ts
@@ -3,7 +3,7 @@ import Elysia, { t } from 'elysia';
 
 import type { Topic } from '#/generated/prisma/client';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
 import { topicNameCache } from '#/lib/cache';
 import { prisma } from '#/lib/prisma';
