+++ packages/emitsignal-server/src/http/webhooks/__tests__/create.spec.ts
@@ -6,7 +6,7 @@ import { prismaMock } from '#/__tests__/mocks';
 mock.module('#/lib/prisma', () => ({ prisma: prismaMock }));
 // Header-driven so a leak into other test files behaves like the real module
 // (no test header → anonymous).
-mock.module('#/http/auth/plugin', () => ({
+mock.module('#/http/auth/resolve-user-id', () => ({
     resolveUserId: ({ headers }: { headers: Record<string, string | undefined> }) =>
         Promise.resolve(headers['x-test-user-id'] ?? null),
 }));
