+++ packages/emitsignal-server/src/http/webhooks/update.ts
@@ -1,6 +1,6 @@
 import Elysia, { t } from 'elysia';
 
-import { resolveUserId } from '#/http/auth/plugin';
+import { resolveUserId } from '#/http/auth/resolve-user-id';
 import {
     validateVerificationBody,
     verificationBodySchema,
